マップ作成を別モジュールに分割
