+++ exif_photo/Reader/exif_reader.py
